# docs/swagger.yaml
definitions:
...
  requestmodel.HelpDeskAnswer:
    properties:
      answer:
        type: string
    required:
    - answer
    type: object
  requestmodel.HelpDeskRequest:
    properties:
      message:
        type: string
      name:
        type: string
      phone_number:
        type: string
      subject:
        type: string
    required:
    - message
    - name
    - phone_number
    - subject
    type: object
...
  response.Response:
    properties:
      error: {}
      message:
        type: string
      result: {}
      status_code:
        type: integer
    type: object
...
info:
  contact:
    email: shahabazsulthan4@gmail.com
    name: 'Creator: Shahabaz Sulthan'
  description: "Laptop Lounge - \U0001F680 Your One-Stop Destination for Ultimate
    Laptop Shopping!\n\U0001F6D2\U0001F4BB Browse, compare, and buy top-notch laptops
    effortlessly.\nPowered by cutting-edge technology, we bring you a seamless shopping
    experience.\n\U0001F31F Dive into the future of laptop shopping with Laptop Lounge!
    \U0001F680\U0001F525"
  title: Laptop Lounge API
paths:
...
  /admin/helpdesk/{requestID}:
    patch:
      consumes:
      - application/json
      description: Update the answer for a help desk request.
      parameters:
      - description: ID of the request to update
        in: path
        name: requestID
        required: true
        type: integer
      - description: Updated answer details
        in: body
        name: answer
        required: true
        schema:
          $ref: '#/definitions/requestmodel.HelpDeskAnswer'
      produces:
      - application/json
      responses:
        "200":
          description: Answer updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request. Invalid input.
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error.
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerTokenAuth: []
      summary: Update Help Desk Answer
      tags:
      - Admin Help Desk
...
  /helpdesk/:
    post:
      consumes:
      - application/json
      description: Create a new help desk request.
      parameters:
      - description: Help desk request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requestmodel.HelpDeskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Request created successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request. Invalid input.
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error.
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create Help Desk Request
      tags:
      - Help Desk
  /helpdesk/replayed:
    get:
      consumes:
      - application/json
      description: Retrieve all help desk requests that have been replied to.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved replied requests
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request. Unable to retrieve replied requests.
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Replied Help Desk Requests
      tags:
      - Help Desk
  /helpdesk/unreplayed:
    get:
      consumes:
      - application/json
      description: Retrieve all help desk requests that are yet to be replied to.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved unreplied requests
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Unreplied Help Desk Requests
      tags:
      - Help Desk
...
securityDefinitions:
  BearerTokenAuth:
    description: 'Use your Bearer token for authentication. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  RefreshTokenAuth:
    description: 'Use your Refresh token to obtain a new Bearer token. Example: "{token}"'
    in: header
    name: Refreshtoken
    type: apiKey
swagger: "2.0"